Moved utils to util directory

+++ pkg/mqt-thrift.PKGBUILD
@@ -1,10 +1,16 @@
 # Contributor: Byron Clark <byron@theclarkfamily.name>
-# Contributor: Serge Aleynikov
+# Maintainer: Serge Aleynikov
 # based on thrift-git PKGBUILD
-toolset=${TOOLCHAIN:-gcc}
-TOOLSET=$(tr '[:lower:]' '[:upper:]' <<< ${toolset})
+
+# If TOOLCHAIN env var is set, then the package name
+# will contain "-${toolchain}" suffix in lower case
+# otherwise, it'll end with "-gcc"
+_toolset=$(tr '[:upper:]' '[:lower:]' <<< ${TOOLCHAIN:-gcc})
+TOOLSET=$(tr  '[:lower:]' '[:upper:]' <<< ${_toolset})
+
+pkgsfx=-${_toolset}
 pkgbase=thrift
-pkgname=mqt-thrift-${toolset}
+pkgname=mqt-${pkgbase}${pkgsfx}
 pkgver=0.9.1
 pkgrel=1
 pkgdesc="Scalable cross-language services framework for IPC/RPC"
@@ -12,7 +18,7 @@ arch=(x86_64)
 url="http://thrift.apache.org/"
 license=(Apache2.0)
 #makedepends=(boost java-environment apache-ant apache-ant-maven-tasks python2 php perl perl-bit-vector perl-class-accessor glib2)
-makedepends=(mqt-boost-gcc python2 glib2)
+makedepends=(mqt-boost${pkgsfx} python2 glib2)
 optdepends=('python2: to use Python bindings'
             'erlang: to use Erlang bindings'
             'perl: to use Perl bindings'
@@ -21,7 +27,7 @@ options=(!emptydirs !makeflags)
 source=("https://github.com/saleyn/thrift/archive/mqt-${pkgver}.zip")
 md5sums=('5ee6b49f8b3e37a71d86dc7541c1aa04')
 
-install=${pkgname}.install
+install=mqt-${pkgbase}.install
 
 ENV_DIR=/opt/env/prod
 INSTALL_DIR=/opt/pkg
@@ -90,7 +96,7 @@ package() {
   ln -s ${pkgver} current
 
   cd ${pkgver}
-  ln -s ${TOOLSET}/bin bin
+  [ -L bin ] || ln -s ${TOOLSET}/bin bin
 
 }
